4.1.8 Рефакторинг модуля со вспомогательными функциями
