+++ src/mock/offer-mock.js
@@ -1,4 +1,4 @@
-import { getRandomArrayElement } from '../utils.js';
+import { getRandomArrayElement } from '../view/utils/common.js';
 
 const mockOffers = [
   {
