+++ src/view/utils/common.js
@@ -1,0 +1,9 @@
+function getRandomArrayElement(items) {
+  return items[Math.floor(Math.random() * items.length)];
+}
+
+function getRandomNumber(min, max) {
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
+
+export { getRandomArrayElement, getRandomNumber };
