+++ src/view/utils/list.js
@@ -1,0 +1,7 @@
+import dayjs from 'dayjs';
+
+function humanizeDueDate(dueDate, dateFormat) {
+  return dueDate ? dayjs(dueDate).format(dateFormat) : '';
+}
+
+export { humanizeDueDate };
